% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{BigRNAConnection-class}
\alias{BigRNAConnection-class}
\title{A connection to BigRNA}
\description{
A connection to BigRNA
}
\section{Slots}{

\describe{
\item{\code{url}}{url bigRNA}

\item{\code{bfc}}{The BiocFileCache}
}}

