% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{BigRNAConnection}
\alias{BigRNAConnection}
\title{Create a BigRNAConnection}
\usage{
BigRNAConnection(url = .baseURL, bfc = .get_cache())
}
\arguments{
\item{url}{a character representing the base URL for BigRNA}

\item{bfc}{a \code{\link[BiocFileCache]{BiocFileCache}} object}
}
\description{
Create a BigRNAConnection
}
\examples{
bigrna = BigRNAConnection()
}
