% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{datafile}
\alias{datafile}
\title{Access a data file from BigRNA}
\usage{
datafile(object, path)
}
\arguments{
\item{object}{A BigRNAConnection object}

\item{path}{The "key" for a data file in BigRNA}
}
\description{
Access a data file from BigRNA
}
\seealso{
\code{\link{gqlQuery}}
}
