% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{gqlQuery}
\alias{gqlQuery}
\title{Query BigRNA metadata}
\usage{
gqlQuery(object, query, ...)
}
\arguments{
\item{object}{A BigRNAConnection object}

\item{query}{The query}
}
\description{
Query BigRNA metadata
}
